cleanup
git-svn-id: svn://svn.code.sf.net/p/jsidplay2/code/trunk@2874 05cac4c1-a8e4-46c3-9d4b-80028926d4ef

+++ jsidplay2/src/main/java/builder/netsiddev/NetSIDDev.java
@@ -124,7 +124,7 @@ public class NetSIDDev extends SIDEmu {
 	@Override
 	public void setVoiceMute(int voice, boolean mute) {
 		if (client.getVersion() >= 3 || voice < 3) {
-			client.addAndSend(new Mute((byte) sidNum, (byte) voice, mute));
+			client.addAndSend(new Mute(sidNum, (byte) voice, mute));
 		}
 	}
