+++ jsidplay2/src/main/java/ui/musiccollection/SearchCriteria.java
@@ -69,7 +69,7 @@ public class SearchCriteria<DECLARING_CLASS, JAVA_TYPE> {
 
 	public static String getText(Object value) {
 		if (value instanceof Integer && (int) value > 255) {
-			return String.format("0x%04X (%d)", (Integer) value, (Integer) value);
+			return String.format("0x%04X (%d)", value, value);
 		}
 		if (value instanceof LocalDateTime) {
 			return ((LocalDateTime) value).format(DateTimeFormatter.ISO_LOCAL_DATE);
